chore: move useFocusOutside to std
