+++ packages/ui/src/components/va-dropdown/VaDropdown.vue
@@ -32,7 +32,7 @@ import { useCursorAnchor } from './hooks/useCursorAnchor'
 import { CursorAnchor, DropdownOffsetProp } from './types'
 import { useDropdown } from './hooks/useDropdown'
 import { warn } from '../../utils/console'
-import { useFocusOutside } from '../../composables/useFocusOutside'
+import { useFocusOutside } from '../../composables/std/event/useFocusOutside'
 import { StringWithAutocomplete } from '../../utils/types/prop-type'
 import { useZIndex } from '../../composables/useZIndex'
 import { focusFirstFocusableChild } from '../../utils/focus'
