+++ packages/ui/src/components/va-menu/VaMenu.vue
@@ -15,12 +15,12 @@
 </template>
 
 <script lang="ts">
-import { nextTick, ref } from 'vue'
+import { nextTick, ref, watchEffect } from 'vue'
 import { VaDropdown, VaDropdownContent } from '../va-dropdown'
 import { VaMenuList } from '../va-menu-list'
 import { extractComponentProps, extractComponentEmits, filterComponentProps } from '../../utils/component-options'
-import { useImmediateFocus, useComponentPresetProp } from '../../composables'
-import { focusFirstFocusableChild } from '../../utils/focus'
+import { useComponentPresetProp } from '../../composables'
+import { focusElement, focusFirstFocusableChild } from '../../utils/focus'
 import { unwrapEl } from '../../utils/unwrapEl'
 import { useChildComponents, defineChildProps } from '../../composables/useChildComponents'
 
@@ -55,7 +55,13 @@ const emit = defineEmits([
 const menuList = ref<HTMLElement>()
 const dropdown = ref<typeof VaDropdown>()
 
-useImmediateFocus(menuList)
+watchEffect(() => {
+  if (menuList.value) {
+    nextTick(() => {
+      focusElement(unwrapEl(menuList.value))
+    })
+  }
+})
 
 const close = () => {
   dropdown.value?.hide()
