+++ packages/ui/src/composables/index.ts
@@ -17,7 +17,6 @@ export * from './useElementRef'
 export * from './useEmitProxy'
 export * from './std/event/useEvent'
 export * from './useFixedBar'
-export * from './useFocusDeep'
 export * from './useFormField'
 export * from './useForm'
 export * from './useHover'
@@ -63,7 +62,6 @@ export * from './useDropdownable'
 export * from './std/internal/useTeleported'
 export * from './useElementTextColor'
 export * from './std/browser/useElementBackground'
-export * from './useImmediateFocus'
 export * from './useNumericProp'
 export * from './useInputMask'
 export * from './std/browser/useElementRect'
@@ -76,3 +74,4 @@ export * from './fabrics'
 export * from './std/internal/useMount'
 export * from './std/browser/useElementScrollableParent'
 export * from './std/browser/useElementFocused'
+export * from './std/browser/useElementFocusedWithin'
