+++ packages/ui/src/composables/std/event/useFocusOutside.ts
@@ -1,29 +1,9 @@
-import { Ref, unref } from 'vue'
+import { MaybeRef, Ref, unref } from 'vue'
 
-import { useEvent } from './'
-import { unwrapEl } from '../utils/unwrapEl'
-
-const checkIfElementChild = (
-  parent: HTMLElement,
-  child: HTMLElement | Window | null | undefined,
-): boolean => {
-  if (!child) {
-    return false
-  }
-  if (child instanceof Window) {
-    return false
-  }
-  if (child.parentElement === parent) {
-    return true
-  }
-
-  return parent.contains(child)
-}
-
-type MaybeRef<T> = T | Ref<T>;
-type MaybeArray<T> = T | T[];
-
-const safeArray = <T>(a: MaybeArray<T>) => (Array.isArray(a) ? a : [a])
+import { useEvent } from './useEvent'
+import { unwrapEl } from '../../../utils/unwrapEl'
+import { MaybeArray, safeArray } from '../../../utils/safe-array'
+import { isElementChild } from '../../../utils/is-element-child'
 
 export const useFocusOutside = (
   elements: MaybeArray<MaybeRef<HTMLElement | undefined>>,
@@ -55,7 +35,7 @@ export const useFocusOutside = (
 
     const isFocusInside = safeArray(elements).some((element) => {
       const el = unwrapEl(unref(element))
-      return el && checkIfElementChild(el, focusTarget)
+      return el && isElementChild(el, focusTarget)
     })
 
     if (!isFocusInside) {
