+++ packages/ui/src/composables/std/index.ts
@@ -15,4 +15,5 @@ export * from './internal/useMount'
 export * from './browser/useElementWidth'
 export * from './browser/useElementScrollableParent'
 export * from './browser/useElementFocused'
+export * from './browser/useElementFocusedWithin'
 export * from './event/useClickOutside'
