+++ packages/ui/src/utils/safe-array.ts
@@ -1,0 +1,3 @@
+export type MaybeArray<T> = T | T[];
+
+export const safeArray = <T>(a: MaybeArray<T>) => (Array.isArray(a) ? a : [a])
